include/fi_abi: Rename fi_abi to ofi_abi
This is an internal header, use the ofi prefix instead of
fi, which is to be reserved for external APIs only.
