+++ src/abi_1_0.c
@@ -38,7 +38,7 @@
 #include <stddef.h>
 
 #include <rdma/fabric.h>
-#include <fi_abi.h>
+#include <ofi_abi.h>
 #include <fi_util.h>
